gh-489 stop switching tabs on add/remove safe; fix action sheet appearance for iPad
