+++ Multisig/UI/Safe Management/Load Safe/AddSafeIntroView.swift
@@ -47,14 +47,11 @@ struct AddSafeIntroView: View {
     }
 
     var loadSafeButton: some View {
-        NavigationLink(destination: EnterSafeAddressView {
-            self.showsLoadSafe.toggle()
-        }, isActive: $showsLoadSafe) {
+        NavigationLink(destination: EnterSafeAddressView(), isActive: $showsLoadSafe) {
             Text("Load Safe Multisig")
         }
         .buttonStyle(GNOFilledButtonStyle())
     }
-
 }
 
 struct AddSafeIntro_Previews: PreviewProvider {
